# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "User subscriptions API",
        "contact": {},
        "version": "1"
    },
    "basePath": "/",
    "paths": {
        "/api/sub": {
            "post": {
                "description": "Create a subscription",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "PostSub",
                "parameters": [
                    {
                        "description": "Structure of new subscription",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/subs.subJSON"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/sub/{id}": {
            "get": {
                "description": "Get a subscription by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "GetSub",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID (int) of specific subscription",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "put": {
                "description": "Update a subscription",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "PutSub",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of subscription",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Structure of subscription",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/subs.subJSON"
                        }
                    }
                ],
                "responses": {}
            },
            "delete": {
                "description": "Delete a subscription",
                "produces": [
                    "application/json"
                ],
                "summary": "DeleteSub",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of subscription",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/subs": {
            "get": {
                "description": "Get all subscriptions",
                "produces": [
                    "application/json"
                ],
                "summary": "GetSubs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID, if need to get all subscriptions of a specific user",
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {}
            },
            "delete": {
                "description": "Delete all subscriptions of specific user",
                "produces": [
                    "application/json"
                ],
                "summary": "DeleteUserSubs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "subs.subJSON": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer"
                },
                "service_name": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        }
    }
}